+++ random_questions.h
@@ -5,6 +5,6 @@
 #include <stdlib.h>
 
 char get_weighted_option(char correct_question, int probability, char *wrong_options, int number_options);
-int get_random_number(int min, int max);
+int get_random_number(unsigned short min, unsigned short max);
 
 #endif
